fix(rebuild): snap rebuild fix
when we are transitioning to AFS mode and we get a snapshot
request from the user, there is a race that if snapshot
command is received after we decided to do AFS transition,
we will miss or have the inconsistant snapshot.

Here traget it self will send prepare snapshot comand to make
each replica ready to receive the snapshot command and then it will
send the snapshot command. And replicas can take the decision as
once this prepare command is successful from all the replica,
then only targe will send the snapshot command. So scanner
can take the decison correctly and can wait for the snapshot command
to finish before go into the AFS mode.

Signed-off-by: Pawan <pawan@mayadata.io>