fix(rebuild): snap rebuild fix (#20)
when zv transitioning to AFS mode and if zv get a snapshot
request from the user, there is a race that if snapshot
command is received after decided to do AFS transition,
snapshot can get missed out or can be in inconsistent state.

Here traget it self will send prepare snapshot comand to make
each replica ready to receive the snapshot command and then it will
send the snapshot command. And replicas can take the decision as
once this prepare command is successful from all the replica,
then only target will send the snapshot command. So scanner
can take the decison correctly and can wait for the snapshot command
to finish before go into the AFS mode.

Signed-off-by: Pawan <pawan@mayadata.io>