Remove our userspace implementation of io_getevents.
This userspace implementation was added because it saves a system call if it
would not return any new events at that point. However, to do this, it has to
cast an intentionally opaque pointer, assume the structure of a kernel object,
and read and write to it, which invokes undefined behaviour if the kernel
structure is actually different. Concretely, the structure indeed seems to be
different for a kernel I am running on arm64.

To ensure no undefined behaviour can happen, remove this optimization.

Signed-off-by: Sjors Gielen <sjors@sjorsgielen.nl>